requirements.txt
git+https://github.com/Pycord-Development/pycord.git
requests
python-dotenv